requirements.txt
livekit-agents==0.15.0
livekit-plugins-silero==0.7.0
livekit-plugins-groq==0.9.0
livekit-plugins-google-generative-ai==0.7.0
livekit-api==0.11.0
python-dotenv==1.0.0
groq==0.7.0
python-multipart==0.0.6
pydantic==2.5.0
aiohttp==3.9.0
